% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorMetaScreener.R
\name{CorMetaScreener}
\alias{CorMetaScreener}
\title{CorMetaScreener: a multi-level meta-analysis framework robustly screens directional signatures based on correlation analyses.}
\usage{
CorMetaScreener(
  expression_accession_vector,
  geneSets_gmt,
  enrichment_method = c("gsva_pearson", "gsva_kendall", "gsva_spearman", "gsva_lm",
    "gsva_biweight", "gsva_distance", "gsva_percentage", "gsva_blomqvist",
    "gsva_hoeffding", "gsva_gamma", "ssgsea_pearson", "ssgsea_kendall",
    "ssgsea_spearman", "ssgsea_lm", "ssgsea_biweight", "ssgsea_distance",
    "ssgsea_percentage", "ssgsea_blomqvist", "ssgsea_hoeffding", "ssgsea_gamma",
    "zscore_pearson", "zscore_kendall", "zscore_spearman", "zscore_lm",
    "zscore_biweight", "zscore_distance", "zscore_percentage", "zscore_blomqvist",
    "zscore_hoeffding", 
     "zscore_gamma", "plage_pearson", "plage_kendall",
    "plage_spearman", "plage_lm", "plage_biweight", "plage_distance", "plage_percentage",
    "plage_blomqvist", "plage_hoeffding", "plage_gamma", "pca_pearson", "pca_kendall",
    "pca_spearman", "pca_lm", "pca_biweight", "pca_distance", "pca_percentage",
    "pca_blomqvist", "pca_hoeffding", "pca_gamma", "aucell_pearson", "aucell_kendall",
    "aucell_spearman", "aucell_lm", "aucell_biweight", "aucell_distance",
    "aucell_percentage", "aucell_blomqvist", "aucell_hoeffding", 
     "aucell_gamma",
    "ucell_pearson", "ucell_kendall", "ucell_spearman", "ucell_lm", "ucell_biweight",
    "ucell_distance", "ucell_percentage", "ucell_blomqvist", "ucell_hoeffding",
    "ucell_gamma", "singscore_pearson", "singscore_kendall", "singscore_spearman",
    "singscore_lm", "singscore_biweight", "singscore_distance", "singscore_percentage",
    "singscore_blomqvist", "singscore_hoeffding", "singscore_gamma", "median_pearson",
    "median_kendall", "median_spearman", "median_lm", "median_biweight",
    "median_distance", 
     "median_percentage", "median_blomqvist", "median_hoeffding",
    "median_gamma", "pearson_fgsea", "kendall_fgsea", "spearman_fgsea", "lm_fgsea",
    "biweight_fgsea", "distance_fgsea", "percentage_fgsea", "blomqvist_fgsea",
    "hoeffding_fgsea", "gamma_fgsea", "pearson_ora", "kendall_ora", "spearman_ora",
    "lm_ora", "biweight_ora", "distance_ora", "percentage_ora", "blomqvist_ora",
    "hoeffding_ora", "gamma_ora", "consensus_pearson", "consensus_kendall",
    "consensus_spearman", "consensus_lm", "consensus_biweight", 
    
    "consensus_distance", "consensus_percentage", "consensus_blomqvist",
    "consensus_hoeffding", "consensus_gamma", "mdt_pearson", "mdt_kendall",
    "mdt_spearman", "mdt_lm", "mdt_biweight", "mdt_distance", "mdt_percentage",
    "mdt_blomqvist", "mdt_hoeffding", "mdt_gamma", "mlm_pearson", "mlm_kendall",
    "mlm_spearman", "mlm_lm", "mlm_biweight", "mlm_distance", "mlm_percentage",
    "mlm_blomqvist", "mlm_hoeffding", "mlm_gamma", "udt_pearson", "udt_kendall",
    "udt_spearman", "udt_lm", "udt_biweight", "udt_distance", 
     "udt_percentage",
    "udt_blomqvist", "udt_hoeffding", "udt_gamma", "ulm_pearson", "ulm_kendall",
    "ulm_spearman", "ulm_lm", "ulm_biweight", "ulm_distance", "ulm_percentage",
    "ulm_blomqvist", "ulm_hoeffding", "ulm_gamma", "viper_pearson", "viper_kendall",
    "viper_spearman", "viper_lm", "viper_biweight", "viper_distance", "viper_percentage",
    "viper_blomqvist", "viper_hoeffding", "viper_gamma", "wmean_pearson",
    "wmean_kendall", "wmean_spearman", "wmean_lm", "wmean_biweight", "wmean_distance",
    "wmean_percentage", 
     "wmean_blomqvist", "wmean_hoeffding", "wmean_gamma",
    "norm_wmean_pearson", "norm_wmean_kendall", "norm_wmean_spearman", "norm_wmean_lm",
    "norm_wmean_biweight", "norm_wmean_distance", "norm_wmean_percentage",
    "norm_wmean_blomqvist", "norm_wmean_hoeffding", "norm_wmean_gamma",
    "corr_wmean_pearson", "corr_wmean_kendall", "corr_wmean_spearman", "corr_wmean_lm",
    "corr_wmean_biweight", "corr_wmean_distance", "corr_wmean_percentage",
    "corr_wmean_blomqvist", "corr_wmean_hoeffding", "corr_wmean_gamma", 
    
    "wsum_pearson", "wsum_kendall", "wsum_spearman", "wsum_lm", "wsum_biweight",
    "wsum_distance", "wsum_percentage", "wsum_blomqvist", "wsum_hoeffding", "wsum_gamma",
    "norm_wsum_pearson", "norm_wsum_kendall", "norm_wsum_spearman", "norm_wsum_lm",
    "norm_wsum_biweight", "norm_wsum_distance", "norm_wsum_percentage",
    "norm_wsum_blomqvist", "norm_wsum_hoeffding", "norm_wsum_gamma", "corr_wsum_pearson",
    "corr_wsum_kendall", "corr_wsum_spearman", "corr_wsum_lm", "corr_wsum_biweight",
    "corr_wsum_distance", 
     "corr_wsum_percentage", "corr_wsum_blomqvist",
    "corr_wsum_hoeffding", "corr_wsum_gamma"),
  min.sz = 1,
  max.sz = 10000,
  p_combine_method = c("geometric_mean", "invt", "invchisq", "logitp", "cct", "meanp",
    "meanz", "sumlog", "sumz", "sump", "votep", "wilkinsonp")
)
}
\arguments{
\item{expression_accession_vector}{A vector of all the sub-dataset names.}

\item{geneSets_gmt}{Genesets to be screened.(Please make sure that every gene set in there is unique, otherwise there will be severe covariance.)}

\item{enrichment_method}{Selection of enrichment methods, multiple options available.}

\item{min.sz}{The genesets contain the minimum number of genes, where the default parameter is 1.}

\item{max.sz}{The genesets contain the maximum number of genes, where the default parameter is 10000.}

\item{p_combine_method}{The choice of method for integrating p-values.}
}
\description{
CorMetaScreener: a multi-level meta-analysis framework robustly screens directional signatures based on correlation analyses.
}
\author{
Dingkang Zhao
}
