% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffMetaScreener.R
\name{DiffMetaScreener}
\alias{DiffMetaScreener}
\title{DiffMetaScreener: a multi-level meta-analysis framework robustly screens directional signatures based on difference analyses.}
\usage{
DiffMetaScreener(
  expression_accession_vector,
  comparisons,
  geneSets_gmt,
  enrichment_method = c("gsva_t", "gsva_limma", "gsva_anova", "gsva_wilcoxon",
    "gsva_permutation", "gsva_kruskal", "ssgsea_t", "ssgsea_limma", "ssgsea_anova",
    "ssgsea_wilcoxon", "ssgsea_permutation", "ssgsea_kruskal", "zscore_t",
    "zscore_limma", "zscore_anova", "zscore_wilcoxon", "zscore_permutation",
    "zscore_kruskal", "plage_t", "plage_limma", "plage_anova", "plage_wilcoxon",
    "plage_permutation", "plage_kruskal", "pca_t", "pca_limma", "pca_anova",
    "pca_wilcoxon", "pca_permutation", "pca_kruskal", "aucell_t", "aucell_limma", 
    
    "aucell_anova", "aucell_wilcoxon", "aucell_permutation", "aucell_kruskal", "ucell_t",
    "ucell_limma", "ucell_anova", "ucell_wilcoxon", "ucell_permutation", "ucell_kruskal",
    "singscore_t", "singscore_limma", "singscore_anova", "singscore_wilcoxon",
    "singscore_permutation", "singscore_kruskal", "median_t", "median_limma",
    "median_anova", "median_wilcoxon", "median_permutation", "median_kruskal", "t_fgsea",
    "limma_fgsea", "anova_fgsea", "wilcoxon_fgsea", "permutation_fgsea", "kruskal_fgsea",
    "t_ora", 
     "limma_ora", "anova_ora", "wilcoxon_ora", "permutation_ora",
    "kruskal_ora", "consensus_t", "consensus_limma", "consensus_anova",
    "consensus_wilcoxon", "consensus_permutation", "consensus_kruskal", "mdt_t",
    "mdt_limma", "mdt_anova", "mdt_wilcoxon", "mdt_permutation", "mdt_kruskal", "mlm_t",
    "mlm_limma", "mlm_anova", "mlm_wilcoxon", "mlm_permutation", "mlm_kruskal", "udt_t",
    "udt_limma", "udt_anova", "udt_wilcoxon", "udt_permutation", "udt_kruskal", "ulm_t",
    "ulm_limma", "ulm_anova", "ulm_wilcoxon", 
     "ulm_permutation", "ulm_kruskal",
    "viper_t", "viper_limma", "viper_anova", "viper_wilcoxon", "viper_permutation",
    "viper_kruskal", "wmean_t", "wmean_limma", "wmean_anova", "wmean_wilcoxon",
    "wmean_permutation", "wmean_kruskal", "norm_wmean_t", "norm_wmean_limma",
    "norm_wmean_anova", "norm_wmean_wilcoxon", "norm_wmean_permutation",
    "norm_wmean_kruskal", "corr_wmean_t", "corr_wmean_limma", "corr_wmean_anova",
    "corr_wmean_wilcoxon", "corr_wmean_permutation", "corr_wmean_kruskal", "wsum_t",
    "wsum_limma", 
     "wsum_anova", "wsum_wilcoxon", "wsum_permutation",
    "wsum_kruskal", "norm_wsum_t", "norm_wsum_limma", "norm_wsum_anova",
    "norm_wsum_wilcoxon", "norm_wsum_permutation", "norm_wsum_kruskal", "corr_wsum_t",
    "corr_wsum_limma", "corr_wsum_anova", "corr_wsum_wilcoxon", "corr_wsum_permutation",
    "corr_wsum_kruskal"),
  min.sz = 1,
  max.sz = 10000,
  p_combine_method = c("geometric_mean", "invt", "invchisq", "logitp", "cct", "meanp",
    "meanz", "sumlog", "sumz", "sump", "votep", "wilkinsonp")
)
}
\arguments{
\item{expression_accession_vector}{A vector of all the sub-dataset names.}

\item{comparisons}{A file name is needed for each sub-dataset that distinguishes between high(activation) and low(inhibition) biological activity. File column 1: Accession, column 2: Group1, column 3: Group2, column 4: Group1_Status, column 5: Group2_Status.}

\item{geneSets_gmt}{Genesets to be screened.(Please make sure that every gene set in there is unique, otherwise there will be severe covariance.)}

\item{enrichment_method}{Selection of enrichment methods, multiple options available.}

\item{min.sz}{The genesets contain the minimum number of genes, where the default parameter is 1.}

\item{max.sz}{The genesets contain the maximum number of genes, where the default parameter is 10000.}

\item{p_combine_method}{The choice of method for integrating p-values.}
}
\description{
DiffMetaScreener: a multi-level meta-analysis framework robustly screens directional signatures based on difference analyses.
}
\author{
Dingkang Zhao
}
